src/main/webapp/app/entities/task-4/delete/task-4-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { Task4Service } from '../service/task-4.service';

import { Task4DeleteDialogComponent } from './task-4-delete-dialog.component';

describe('Task4 Management Delete Component', () => {
  let comp: Task4DeleteDialogComponent;
  let fixture: ComponentFixture<Task4DeleteDialogComponent>;
  let service: Task4Service;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, Task4DeleteDialogComponent],
      providers: [NgbActiveModal],
    })
      .overrideTemplate(Task4DeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(Task4DeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(Task4Service);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
